+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- Added an example of using metric views to customize instruments. (#3177)
+
 ## [0.32.1] Metric SDK (Alpha) - 2022-09-22
 
 ### Changed